python-service/requirementsMin.txt
Flask==3.1.1
flask-cors==6.0.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.6
langchain-text-splitters==0.3.9
numpy==2.2.6
pandas==2.3.1
pillow==11.3.0
python-dotenv==1.1.1
Werkzeug==3.1.3
